Added a team files zipper (not actually zipping yet...need library)

+++ LeaguevinePostingLog.m
@@ -180,7 +180,7 @@
 -(void)switchLogsIfNeeded {
     if ([[NSFileManager defaultManager] fileExistsAtPath:self.currentLogFilePath]) {
         NSDictionary *attributes = [[NSFileManager defaultManager] attributesOfItemAtPath:self.currentLogFilePath error:nil];
-        long size = [attributes fileSize];
+        unsigned long long size = [attributes fileSize];
         if (size > 500000) {
             [self switchLogs];
         }
